refactor(be): modules 폴더로 합치기
