+++ server-toss/src/modules/ad/ad-view.entity.ts
