+++ server-toss/src/modules/drawing/drawing.entity.ts
@@ -6,8 +6,8 @@ import {
   Property,
 } from "@mikro-orm/decorators/legacy";
 import { BaseEntity } from "src/common/base.entity";
-import { Prompt } from "src/prompt/prompt.entity";
-import { User } from "src/user/user.entity";
+import { Prompt } from "src/modules/prompt/prompt.entity";
+import { User } from "src/modules/user/user.entity";
 
 @Entity({ tableName: "drawings" })
 export class Drawing extends BaseEntity {
