+++ server-toss/src/modules/prompt/daily-prompt.entity.ts
