+++ server-toss/src/modules/ranking/ranking.entity.ts
